# Conflicts:
#	target/classes/Main.class

+++ src/main/java/MachineMatierePremiere.java
@@ -1,9 +1,9 @@
 import java.util.HashMap;
 
 public class MachineMatierePremiere extends Machine{
-    public HashMap<MatierePremiere, Integer> capacite;
+    public HashMap<MatierePremiere, Double> capacite;
 
-    public MachineMatierePremiere(String nom,HashMap<MatierePremiere, Integer> capacite) {
+    public MachineMatierePremiere(String nom,HashMap<MatierePremiere, Double> capacite) {
         super(nom);
         this.capacite=capacite;
     }
